Refactor the app for easier extensibility (#2)
This is a wholesale change in approach in how the app is structured, initialized, and runs. The major impetus is to bring extensibility into the forefront. This required not only pulling the various different tags and their functionality into a separate module, but a wholesale rewrite of how the various parts of the app communicate with each other.

Tags being added or removed now dispatch events to a number of event listeners which have been registered. These listeners are encapsulated within plugins which contain all the logic for handling only the events which are important to the plugin. For example, all the logic about playing songs from Spotify, along with the tag object which contains the configuration for a particular physical NFC tag, makes up a plugin; nothing else in the app knows anything about Spotify.

Once the event is handled, the main app runner is notified of the completion by another dispatched event. Typically, this is signified by updating the Dimensions pad color.

Plugins should be registered in musicfig/__init__.py, which will then cause the init_app to register the tag type with NFCTagManager, add event handlers, etc.

Part of this update has deprecated the Slack functionality, opting instead for the more general Webhook plugin which performs exactly the same functionality, only more broadly. Another thing which is missing after this update is the local MP3 playback because, frankly, I won't be using it. I am planning on adding it back in, but need to make some product-level decisions on what functionality to support. Some open questions along that path:

- since the Spotify device could be anything and the MP3 playback comes from the Raspberry Pi audio jack, should we allow the two to play at the same time, e.g. with zones?
- Should multiple MP3s be allowed to play at once, creating a kind of sound board? (is this even possible with the pi hardware?)
- If the answer to the above two is "no", how do we disable playback of the previously-playing thing? Do we allow resume of Spotify if we interrupted with MP3 and vice versa?